Completely re-created Cottontail DB's cost model and query planning algorithm: - Cost of execution is now expressed in terms of IO, CPU and memory which are considered jointly - The new query planner now recursively re-writes the logical plan in order to arrive at the 'optimal' execution plan.

+++ src/main/kotlin/org/vitrivr/cottontail/math/knn/metrics/DistanceKernel.kt
@@ -12,7 +12,7 @@ import org.vitrivr.cottontail.model.values.types.VectorValue
  */
 interface DistanceKernel {
     /** Estimate of the cost required per vector component. */
-    val cost: Double
+    val cost: Float
 
     /**
      * Calculates the distance between two [VectorValue]s.
